refactor(structure): unify domain modules and eliminate single-file folders
Playback engine consolidation:
- Relocate src/lib/playback/frameSkipping.ts → src/core/playback/
- Export frameSkipping from src/core/playback/index.ts
- Remove src/lib/playback/

Monitoring engine consolidation:
- Relocate LifecycleMonitor, PerformanceMonitor, ResourceTracker
  from src/lib/monitoring/ → src/core/monitoring/
- Add barrel export at src/core/monitoring/index.ts
- Update all import sites: App.tsx, textRasterizer.ts,
  PreviewMediaPool.ts, ProjectSession.ts, ProjectStateReset.ts,
  projectStore.ts, ThumbnailWorkerPool.ts
- Remove src/lib/monitoring/

Single-file folder elimination:
- PreviewQualityManager.ts → src/components/editor/preview/
- windowState.ts          → src/lib/platform/
- calculator.ts           → src/components/editor/transform/
- sequenceAutoAspect.ts   → src/lib/timeline/
- Remove empty dirs: lib/preview, lib/window, lib/transform, lib/sequence

Update all affected import sites across hooks, stores, and components.

TypeScript: 0 errors  |  Rust: 0 warnings  |  No behaviour change.
