+++ src/components/editor/preview/PreviewQualityManager.ts
