+++ src/components/editor/preview/SourcePreview.tsx
@@ -7,7 +7,7 @@ import { getInsertIndexForNewTrack, useTimelineStore } from "@/store/timelineSto
 import { useProjectStore } from "@/store/projectStore";
 import { createClipFromAsset } from "@/lib/timeline/timelineClip";
 import { getActiveSessionOrNull } from "@/core/runtime/ProjectSession";
-import { autoAdaptSequenceForFirstVisualClip } from "@/lib/sequence/sequenceAutoAspect";
+import { autoAdaptSequenceForFirstVisualClip } from "@/lib/timeline/sequenceAutoAspect";
 import { DEFAULT_PLACEMENT_POLICY, resolveAddToTimelinePlacement, resolveDefaultFitModeForAsset } from "@/lib/timeline/placementPolicy";
 import { getPlaybackClock } from "@/hooks/usePlaybackClock";
 import type { SourcePlaybackContext } from "@/core/playback";
