+++ src/core/monitoring/LifecycleMonitor.ts
