+++ src/core/monitoring/PerformanceMonitor.ts
