+++ src/core/resources/PreviewMediaPool.ts
@@ -32,8 +32,8 @@ import type { Clip, MediaAsset, TransitionTimelineItem } from "@/types";
 import { convertFileSrc } from "@tauri-apps/api/core";
 import { isWebviewOrExternalUrl } from "@/lib/platform/pathConversion";
 import { resolveClipSourceTime } from "../timeline/sourceTime";
-import { performanceMonitor } from "@/lib/monitoring/PerformanceMonitor";
-import { resourceTracker } from "@/lib/monitoring/ResourceTracker";
+import { performanceMonitor } from "@/core/monitoring/PerformanceMonitor";
+import { resourceTracker } from "@/core/monitoring/ResourceTracker";
 import { useTimelineStore } from "../../store/timelineStore";
 import { PreviewPlaybackScheduler, type MediaAction, type MediaElementState } from "../playback/PreviewPlaybackScheduler";
 import { VideoTextureManager } from "../render/VideoTextureManager";
