+++ src/lib/platform/windowState.ts
