+++ src/lib/timeline/sequenceAutoAspect.ts
